requirements.txt
Flask
flask-cors
openai-whisper
torch
yt-dlp
requests
